Rename package names to use all lowercase
