--- src/main/java/Duke/Main.java
@@ -1,12 +1,0 @@
-package Duke;
-
-import javafx.application.Application;
-
-/**
- * Main entry point to the program.
- */
-public class Main {
-    public static void main(String[] args) {
-        Application.launch(Duke.class, args);
-    }
-}
