--- src/main/java/Duke/Task/InvalidTaskException.java
@@ -1,15 +1,0 @@
-package Duke.Task;
-
-import Duke.DukeException;
-
-public class InvalidTaskException extends DukeException {
-    private static final String INVALID_TASK_MESSAGE = "Invalid task: %s";
-
-    public InvalidTaskException(String message) {
-        super(String.format(INVALID_TASK_MESSAGE, message));
-    }
-
-    public InvalidTaskException(String message, Throwable cause) {
-        super(String.format(INVALID_TASK_MESSAGE, message), cause);
-    }
-}
