+++ src/main/java/duke/Main.java
@@ -1,0 +1,12 @@
+package duke;
+
+import javafx.application.Application;
+
+/**
+ * Main entry point to the program.
+ */
+public class Main {
+    public static void main(String[] args) {
+        Application.launch(Duke.class, args);
+    }
+}
