+++ src/main/java/duke/ui/controllers/MainWindow.java
@@ -1,0 +1,50 @@
+package duke.ui.controllers;
+
+import duke.Duke;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+
+/**
+ * Controller for Main Window.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Duke duke;
+
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+    }
+
+    public void setDuke(Duke d) {
+        duke = d;
+    }
+
+    /**
+     * Handles a user input event.
+     * Echos the input, then parses and runs it.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        dialogContainer.getChildren().addAll(DialogBox.ofUser(input));
+        duke.parseAndRun(input);
+        userInput.clear();
+    }
+
+    public void addDukeMessage(String message) {
+        dialogContainer.getChildren().addAll(DialogBox.ofDuke(message));
+    }
+}
