+++ src/main/resources/view/MainWindow.fxml
@@ -9,7 +9,7 @@
 
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
             xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="Duke.Ui.Controllers.MainWindow">
+            fx:controller="duke.ui.controllers.MainWindow">
     <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
                promptText="Enter command..."
                style="-fx-background-color: #ecf0f1; -fx-font-family: monospace; -fx-background-radius: 0; -fx-font-weight: 500;"
